Refactor to pytorch-lightning CLI
Adapt a similar scheme as the bone-disorder project whereby everything is controlled by the ptl CLI.
Implement new network types for explicit multi-task sex prediction.
Store prediction normalizations in the model checkpoints.

+++ lib/modules/metrics.py
@@ -4,7 +4,7 @@ from torch import Tensor
 
 
 class RescaledMAE(torchmetrics.MeanAbsoluteError):
-    def __init__(self, compute_on_step: Optional[bool] = None, rescale: int = 1):
+    def __init__(self, compute_on_step: Optional[bool] = None, rescale: float = 1):
         super().__init__(compute_on_step=compute_on_step)
         self.rescale = rescale
